API queries database
live data is reflected in the Statistic Selection portion of the website. Everything was pair programmed by Sarah and Hayden.

Note: for the error about "ERR_HTTP_HEADERS_SENT: Cannot set headers after they are sent to the client", ensure there is only 1 return. This is a good stackoverflow reference: https://stackoverflow.com/questions/52122272/err-http-headers-sent-cannot-set-headers-after-they-are-sent-to-the-client