Continue compatibility changes
Added in a delay for Chrome to be able to catch the right theme when switching between themes. Did some more work on themes and styling for compatibility cross browsers. These changes are subject to more change later. This branch has lived on too long as it is, and there is too much in here.